created a reusable icon button to be used in the cards
