+++ src/app/models/User.model.ts
@@ -1,6 +1,6 @@
 export class User {
 
-  id!: number;
+  _id!: string;
   name!: string;
   email!: string;
   role!: string;
