+++ src/app/shared-components/buttons/button-card/button-card.component.html
@@ -1,0 +1,3 @@
+<div class="button">
+  <mat-icon class="icon" [fontIcon]="icon" (click)="onButtonClick($event)"></mat-icon>
+</div>
