+++ src/app/shared-components/cards/user-card/user-card.component.css
