+++ src/app/shared-components/cards/user-card/user-card.component.html
@@ -4,7 +4,7 @@
   </section>
   <section class="cardContent">
     <p class="infoLabel">{{ info }}</p>
-    <ButtonCard handlerOnClick={handlerOnClick} />
+    <button-card icon="Person Off" (onClick)="selectUser()"></button-card>
     <section class="name">
       <label class="label">{{ name }}</label>
       <p class="property">{{ user['name'] }}</p>
